Move old files in to different folders for reference during re-design.
