+++ backend/helpers-old/passport.js
@@ -2,7 +2,7 @@ const passport = require('passport');
 const passportJWT = require('passport-jwt');
 const JWTStrategy = passportJWT.Strategy;
 const ExtractJWT = passportJWT.ExtractJwt;
-const User = require('../models/user');
+const User = require('../models-old/user');
 const config = require('../config');
 
 passport.use(new JWTStrategy({
